+++ AuthPermissions/AuthSetupData.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) 2021 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
+using System.Collections.Generic;
+using AuthPermissions.SetupCode;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace AuthPermissions